requirements.txt
astroid==2.0.4
autopep8==1.4
Django==2.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pyinotify==0.9.6
pylint==2.1.1
pynotify==0.1.1
pytz==2018.5
rope==0.11.0
six==1.11.0
wrapt==1.10.11
